fixed shader file bugs causing issues;

+++ unify/src/graphics/opengl.cpp
@@ -57,7 +57,7 @@ namespace unify::graphics::opengl {
 	}
 
 	Texture::Texture(const char* filepath) : m_RendererID(0), m_Filepath(filepath), m_LocalBuffer(NULL), m_Width(0), m_Height(0), m_BPP(0) {
-		
+
 		stbi_set_flip_vertically_on_load(1);
 		m_LocalBuffer = stbi_load(filepath, &m_Width, &m_Height, &m_BPP, 4);
 
@@ -70,7 +70,7 @@ namespace unify::graphics::opengl {
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
 
 		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA8, m_Width, m_Height, 0, GL_RGBA, GL_UNSIGNED_BYTE, m_LocalBuffer);
-		
+
 		glBindTexture(GL_TEXTURE_2D, 0);
 
 		if (m_LocalBuffer) {
@@ -183,11 +183,11 @@ namespace unify::graphics::opengl {
 	}
 
 	void VertexArray::AddBuffer(const VertexBuffer& vBuffer, const VertexBufferLayout& vLayout)
-	{	
+	{
 		Bind();
 		vBuffer.Bind();
 		const auto& elements = vLayout.GetElements();
-		GLuint offset = 0;
+		uintptr_t offset = 0;
 		for (GLuint i = 0; i < elements.size(); i++) {
 			const auto& element = elements[i];
 			glEnableVertexAttribArray(i);
